+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added --force option to finish command,
  which tries to change paused groups with active units directly to Finishing on fah 8.4+.
  This may break in the future.
+- Added short option -f for --force
 
 ---